+++ src/components/common/table/index.js
@@ -1,0 +1,6 @@
+
+import Table from "./table";
+import TableHeader from "./tableHeader";
+import TableBody from "./tableBody";
+export default Table;
+export { TableHeader, TableBody };
