+++ src/components/page/userListPage/index.js
@@ -1,0 +1,2 @@
+import UsersListPage from "./usersListPage";
+export default UsersListPage;
