+++ src/components/ui/loginForm.jsx
@@ -1,0 +1,89 @@
+import React, { useState, useEffect } from "react";
+import { validator } from "../../utils/validator";
+import TextField from "../common/form/textField";
+
+const LoginForm = () => {
+    const [data, setData] = useState({ email: "", password: "" });
+    const [errors, setErrors] = useState({});
+
+    const handleChange = ({ target }) => {
+        setData((prevState) => ({
+            ...prevState,
+            [target.name]: target.value
+        }));
+    };
+
+    const validatorConfig = {
+        email: {
+            isRequired: {
+                message: "Электронная почта обязательна для заполнения"
+            },
+            isEmail: {
+                message: "Email введен не корректно"
+            }
+        },
+        password: {
+            isRequired: {
+                message: "Пароль обязателен для заполнения"
+            },
+            isCapitalSymbol: {
+                message: "Пароль должен содержать хотябы одну заглавную букву"
+            },
+            isContainDigit: {
+                message: "Пароль должен содержать хотябы одну цифру"
+            },
+            min: {
+                message: "Пароль должен содержать не менее 8 символов",
+                value: 8
+            }
+        }
+    };
+
+    useEffect(() => {
+        validate();
+    }, [data]);
+
+    const validate = () => {
+        const errors = validator(data, validatorConfig);
+
+        setErrors(errors);
+        return Object.keys(errors).length === 0;
+    };
+    const isValid = Object.keys(errors).length === 0;
+
+    const handleSubmit = (e) => {
+        e.preventDefault();
+        const isValid = validate();
+        if (!isValid) return;
+        console.log(data);
+    };
+    return (
+        <div className="container mt-5 shadow rounded-3 bg-light bg-gradient card w-50">
+            <div className="row">
+                <div className="col-md-8 offset-md-2 p-4">
+                    <h3 className="mb-2 text-center">Login</h3>
+                    <form onSubmit={handleSubmit}>
+                        <TextField
+                            label="Электронная почта"
+                            name="email"
+                            value={data.email}
+                            onChange={handleChange}
+                            error={errors.email}
+                        />
+                        <TextField
+                            label="Пароль"
+                            type="password"
+                            name="password"
+                            value={data.password}
+                            onChange={handleChange}
+                            error={errors.password}
+                        />
+                        <button type="submit" disabled={!isValid} className="btn btn-primary w-100 mx-auto mt-4">Войти</button>
+                    </form>
+                </div>
+            </div>
+        </div>
+    );
+};
+
+export default LoginForm;
