+++ src/components/ui/qualities/index.js
@@ -1,0 +1,2 @@
+import Qualities from "./qualitiesList";
+export default Qualities;
