+++ src/components/ui/qualities/qualitie.jsx
