+++ src/components/ui/usersTable.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 import PropTypes from "prop-types";
-import Bookmark from "./bookmark";
-import QualitiesList from "./qualitiesList";
-import Table from "./table";
+import Bookmark from "../common/bookmark";
+import Qualities from "./qualities";
+import Table from "../common/table";
 import { Link } from "react-router-dom";
 
 const UsersTable = ({ users, onSort, selectedSort, onToggleBookMark, onDelete, ...rest }) => {
@@ -17,7 +17,7 @@ const UsersTable = ({ users, onSort, selectedSort, onToggleBookMark, onDelete, .
         qualities: {
             name: "Качества",
             component: (user) => (
-                < QualitiesList qualities={user.qualities} />
+                < Qualities qualities={user.qualities} />
             )
         },
         professions: { path: "profession.name", name: "Профессия" },
