+++ src/layouts/users.jsx
@@ -1,12 +1,13 @@
 import React from "react";
 import { useParams } from "react-router-dom";
-import UserInfo from "../components/userInfo";
-import UsersList from "../components/usersList";
+
+import UserPage from "../components/page/userPage";
+import UsersListPage from "../components/page/userListPage";
 
 const Users = () => {
     const params = useParams();
     const { userId } = params;
-    return <>{userId ? <UserInfo userId={userId} /> : <UsersList/>}</>;
+    return <>{userId ? <UserPage userId={userId} /> : <UsersListPage/>}</>;
 };
 
 export default Users;
